encode: introduce EncodeBuilder::apply_to method
base58 is often used for inputs, e.g. Bitcoin addresses or 32-byte
hashes, which can use small statically-sized on-stack buffers when
encoding. On the other hand, utility code which does not know what
kind of input it’s going to deal with cannot assume a input length
limit.  This usually leads to encoding into a Vec or a String even
if the caller doesn’t actually need to own the buffer with encoded
data.

For those situations, introduce EncodeBuilder::apply_to method
which checks the maximum length of the output and if it’s small
enough performs encoding onto the stack.  The limit on the output
length is set such that a 64-byte buffer plus version and checuskm
can be handled without any heap allocation.

With this change, also refine estimation of the output size.  Instead
of using 8/5 ratio (i.e. 1.6) use 1.5 which is closer to the actual
1.37 and also produces simpler expression.